// gemip-mipcittadinowcl/src/app/core/models/incontroPreaccoglienza.ts
/**
 * @license
 *
 * Copyright © 2025 Regione Piemonte
 *
 * Licensed under the EUPL, Version 1.2 or – as soon they will be
 * approved by the European Commission - subsequent versions of the
 * EUPL (the "Licence");
 *
 * You may not use this work except in compliance with the Licence.
 * You may obtain a copy of the Licence at:
 *
 * https://interoperable-europe.ec.europa.eu/collection/eupl/eupl-text-eupl-12
 * or
 * https://opensource.org/license/EUPL-1.2
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the Licence is distributed on an "AS IS" basis,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the Licence for the specific language governing permissions and
 * limitations under the Licence.
 */

/**
 * API mipcommonapi
 * TODO add a description here
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LuogoIncontro } from './luogoIncontro';
import { Operatore } from './operatore';

export interface IncontroPreaccoglienza {
    id?: number;
    denominazione?: string;
    flgIncontroErogatoDaRemoto?: string;
    linkIncontroRemoto?: string;
    luogoIncontro?: LuogoIncontro;
    operatoreCreazione?: Operatore;
    numMaxPartecipanti?: number;
    numPartecipantiIscritti?: number;
    dataIncontro?: Date;
    note?: string;
    codUserInserim?: string;
    dataInserim?: Date;
    codUserAggiorn?: string;
    dataAggiorn?: Date;
}